requirements.txt
google-cloud-vision==1.0.0
google-cloud-texttospeech==2.2.0
picamera==1.13